Applying code patches for zfighting-fix from gitlab source. This patch (#25)
includes addition of a reversed depth buffer in order to resolve
zfighting artifacts caused by a mis-distribution of precision.
The changes implemented to resovle this are as follows:
- Flipped Znear and Zfar planes to push the depth range from 1..0
- Flipped depth test to VK_COMPARE_OP_GREATER
- Flipped depth buffer clear values from 1.0 to 0.0

This patch also includes a change to hoist the glm includes into a
separate glm_common.h. This is to resolve platform issues relating to
inconsistent compilation macros between different include locations.