Adding Vulkan-style statistics counters where available
Previously, the 'stats' interface just used HWCPipe to
interrogate performance counters from the device. HWCPipe
is not widely supported, so this really only worked for
Arm devices.

This change allows multiple 'StatsProviders' to be used.
Each provider can supply all, some, or none of the requested
counters depending on its capabilities. Any counters not
supplied by one provider will be requested of the next in
the list until one is found, or we run out of providers.
So, CPU counters can come from HWCPipe for example, whilst
GPU counters come from Vulkan's extension which has been
added as another provider.

The StatsProvider defines an abstract interface that is
implemented by the concrete providers. The parameters used
for graphing the statistics have moved from the GUI classes
into the StatsProvider also. This allows providers to modify
the names, units and scaling of counters where these differ
from the default values.