Reworked top level stats interface and construction
Based on private feedback from Zandro, I've moved the
construction of the stats object into the base
VulkanSamples class which leaves individual samples
to just request the samples they would like.

The active_frame_index is no longer passed to
update_stats() or the command_buffer_begun/ended()
methods. It is fetched internally by the providers
that require it.