Make physical device selection a bit smarter.
When selecting the physical device, try to select CPU deviceType last. Do so, by
ordering devices in somehow sensible way, at least for graphics work.

It could be even more smart, by returning ordered list, which can be then checked
against the queue and swapchain / present support. But lets not complicate things
too much for a moment.

The motivation, is some users of Mesa, when compile Intel (anv) and Gallium
swrast (lavapipe) Vulkan drivers, Vulkan driver will load them in random order,
and because `anv` devices usually are integrated, the current code will usually
select `lavapipe`, which is a (multi-threaded) CPU implementation, which does
work, but is not what user wants in 99.9% of cases, until it is selected
explicitly.