Dynamic rendering (#394)
* Using basic raytracing as template

* Load sponza

* Loading sponza model

* Separating out ray tracing command buffers

* Add render modes

* Edit vertex structure

* Put static objects in buffer with device-exclusive memory

* Adding vertex and index data to bindings

* Demonstration of referencing model data from ray tracing shader

* Implemented texture array and visualization

* Added fire pit

* Added ray-traced shadows

* Adding support for AO

* Updated model settings

* Fixed a few warnings and spelling errors.

Also fixed the fences problem that causes crash bug from fences on close.

* Separating out static model loading

* Separate handling of static / dynamic buffers

* Added dynamic object and refactored

* Fixed another warning and fixed the fences.

* Switch to using constants and specializations in order to change values.  Can switch back to uniforms at a later date but the uniforms were proving a render issue for linux.  This should also demonstrate how to use multiple specialization constants in a ray traced environment.

* more warnings removal

* Merge warn_remove_and_fence_fix

* Moving instance buffer creation to TLAS construction

* Adding refraction placeholder

* Perform BLAS update for dynamic objects and TLAS update every frame.

* Adding dynamic sinusoidal wave

* Finished refractive wave

* Update to refraction

* Adding separate flame instance

* Working flame / particle generator

* Updates to flame and refraction calculation

* Updated refraction algorithm

* Added texture

* more warnings removal
move the png to ktx format

* ensure we're pointing to the right place

* Merge and update

* Update to ktx commit

* run clang-format over the sample.

* Added transformations

* Merged

* Merged

* Adding transformations

* run a profile pass.
move pts, uvs, normals, indices to be created/initialized once to avoid dynamic memory generation.
remove warnings.
Update copyright.

* Update copyright in CMakeLists.txt
Update sample cmake function definition
Update samples list in Samples CMakeLists.txt
Update README.md with first pass definition of the sample.

* Added tutorial rough draft

* grammar check.

* grammar check.

* Added fence reset

* Updating triangle indexing code

* Avoid dynamic allocation / copy

* update copyright

* Fixes for Clang-format and copyright headers CI failures.

* add in nonuniformEXT to the texture lookup as imageOffset isn't uniform.

* switch to textureLod from texture to ensure it's ready for mipmaps if copied from sample.

* Fixing refraction bug

* fix typo in README.md

* fix typo in README.md and in CMakeLists.txt

* Updates for PR comments

* Adding memory barriers

* Updating shaders to use defines

* clang format cleanup

* add missing copyright to new header

* fix attribution in the tutorial.md

* framework:
acceleration structure: (build) fix a bug which only would allow a single triangle to be included in a acceleration structure.
acceleration structure: (build) allow for more than one call to build (useful in animation).  This was caused by recreating the buffer each time rather than reusing it if it already exists.
acceleration structure: (add triangles) allow for developer to specify the data address of the vertex index and transform buffers that are used.  This allows for multi model gltf files to work as in the sponza scene.
acceleration structure: enable buffer retrieval to allow for developer to add in memory barriers.
acceleration structure: provide developer means to wipe the geometries added to the acceleration structure.  This is very important if animating and thus recalling the same structure(s).  NB: a future feature would be to index the geometries and instances to allow for updating existing data individually rather than a requirement to wipe out everything and rebuild everything.
scratch buffer: add a return for the scratch buffer's size to allow for reusing the scratch buffer if it's the same size as previously allocated.

Sample:
As this is the first ray tracing sample to make use of @Sascha William's excellent acceleration structure framework, and we already have the existing long form, I opted to enable a macro to turn on/off using the framework's acceleration structure.  I think this is valuable to end users to show a precise "what the frameork does" behind the scenes.  Future samples should opt to use the framework.

* run clang-format over scratch_buffer.h

* Add in proper update logic for the acceleration_structure in both the framework and the sample.

* Ensure that only updated geometry are used in each update pass and make framework the default AS.

* Small changes for buffer.

* run clang-format

* fix android

* fix android

* address the returns doc.

* Adding outline for  VK_KHR_dynamic_rendering sample

* Updating dynamic rendering sample

* Adding destructors

* Adding transition layout

* fix crash on device not supported.

* Updating order of prepare

* Updating dynamic rendering sample

* warning removal and validation fixes.

* Removing unused color attachment and updating dynamic load

* Removing unused UV param in vertex shader

* Adding UI

* fix the sampler not getting destroyed.

* clang-format, add the cmake requirements, and add basic README.md text.

* clang-format, fix copyrights

* fix missing copyright

* Adding dynamic rendering tutorial

* Memory barrier

* run clang-format

* Updating image memory barrier per PR comments

* run clang-format

* Update per PR comments

* Updating dst / src stage masks per PR comments

* Updating dst / src stage masks per PR comments

* run clang-format and update from upstream.

* Update dynamic_rendering_tutorial.md

fix spelling.

Co-authored-by: William Lohry <wflohry@wflohry.com>