Auto enable vkb enable portability (#497)
* Fix issue #452

* Attempt some logic to make setting of VKB_ENABLE_PORTABILITY happen if and only if Vulkan_VERSION is at least 1.3 which should include the required portability extension which will enable the instance to find the non-conformant vulkan for moltenVK.

* update copyright.

* make the compile definition explicit.

* Validated as working with Vulkan 1.3.216.0 on MacBook Air M2.

* enable vulkan.hpp samples to work.

* don't crash open_gl_interop.cpp if a gl_context doesn't get created.  As in the case when extensions required are not present.

* revert the 3.19 cmake version upgrade for platforms other than OSX.
NB: This *might* mean that Android on OSX would have an issue due to the version of CMake that Android SDK might ship with is 3.18.  Revisit this once most people are on Vulkan 1.3 and/or confirmation that android on OSX is a problem.

* revert the 3.19 cmake version upgrade for platforms other than OSX.
NB: This *might* mean that Android on OSX would have an issue due to the version of CMake that Android SDK might ship with is 3.18.  Revisit this once most people are on Vulkan 1.3 and/or confirmation that android on OSX is a problem.

* update the Copyright

* update with clang-format

* oddly running clang-format locally resulted in some white space changes which don't agree with the CICD white space results.