Update and standardise sRGB usage (#518)
* Update and standardise sRGB usage

Most displays expect to receive sRGB encoded data.
This change ensures that all of the samples request
sRGB swapchain images. The GPU hardware will then
ensure that the linear result of the fragment shader
is automatically converted to non-linear sRGB when
written to the swapchain.

The performance samples were already selecting sRGB
as the priority format, but the API samples were not.

I've removed the UNORM formats from the swapchain
format priority lists as it doesn't make much sense
to have a UNORM swapchain image unless you want to
manually convert to sRGB at the end of your fragment
shaders - not a sensible approach.

I then checked all of the texture data formats.
Textures that are in UNORM formats are treated as
being linearly encoded by the GPU and enter the
shaders unaltered. SRGB format textures are converted
by the GPU into linear space before being seen by
shaders. This ensures that the shaders are doing
computations on linear data at all times.

Most image authoring tools will write the data
encoded in sRGB. Some file formats don't record
a colorspace though. In those cases the code now
assumes that the data is sRGB encoded, if and only
if that image is going to be used as a color texture.
For image formats that know their colorspace, this
is preserved.

The gltf loader has been modified slightly. The gltf
spec is very clear about which images must be encoded
as sRGB and which have to be linear. That knowledge
is now in the loader, so the texture images are given
the correct format during loading.

I've modified some of the clear colors in the API
samples to convert them into linear space. Previously
they were effectively in sRGB space since the swapchain
was unorm. This ensures that the clear color looks close
to what it did previously.

To summarise, the samples should now be doing this:

For textures that contain color data:
sRGB texture -> GPU-conv -> linear shading -> GPU-conv -> sRGB swapchain

For textures that contain lineaer data (e.g. normals):
unorm texture -> linear shading -> GPU-conv -> sRGB swapchain

* Add assert for non-srgb gltf names and copy comment to hpp version

* Correct clear colors and adjust terrain_tessellation layers

* Updated missed copyright header

* Add comment on maybe_coerce_to_srgb()