direct-to-display window improvements (#513)
* direct-to-display window improvements

Rather than just taking the first plane of the first display
we now gather a list of valid candidates and pick the best
fit.

An additional display 'FullscreenStretch' mode has been added.
This uses VkDisplayPresentInfoKHR from the VK_KHR_display_swapchain
extension, when available, to specify separate src and dst
rectangles for the surface and display. This enables you to
have a 640x360 surface stretched fullscreen onto the display
for example, rather than just appearing in a corner.

For direct-to-display specifying --width & --height controls
the surface size in standard window mode and in stretch mode.
In fullscreen or borderless modes, the surface is the same
size as the display regardless of the width and height settings.

* Small code improvements

From code review comments

* Fix clang format

* Fix error in hpp_window.h

We can't just call into the base class directly
since we've actually been reinterpret_cast from a
derived vkb::Window class. Ensure the derived
class method gets called.

* Fix clang formatting