Reduce memory footprint during gltf loading
The gltf loader was loading all the image data and then
attempting to upload that via staging buffers to images
all at once. That required device allocations for all the
staging buffers and the destination images to exist at
the same time. On small memory systems that easily blew
the budget when running descriptor_management (Bonza4x)
for example.

This change does the uploads in batches of ~64MB of image
data. The staging buffers only need to cover that 64MB
at any one time in that case.