Reduce memory usage for OneAllocationPerBuffer
OneAllocationPerBuffer is used by the decriptor_management
sample. It's intended to show the difference between having
one buffer with one allocation, and having multiple allocations
per buffer.

The way it was written however, each buffer was a minimum of
256KB regardless of the number of allocations. This sample was
making thousands of small allocations, each of which was then
trying to get its own 256KB buffer block. This quickly led to
memory exhaustion on small memory systems.

This change ensures that buffer blocks are allocated only at
thesize requested when OneAllocationPerBuffer is active.