Batch mode improvements
Now logs the name of the sample being started.
This is very useful if a sample crashes in batch mode
as you can easily see from the log which sample it was.

Fixed the wrap-to-start command argument. There is no
.as<bool> on the parser, so it always returned false.
The logic for the wrapping was also slightly wrong in
that the first sample would be missed during the wrap.