Fix validation errors about using unsupported linear filtering (#924)
* Resolve validation errors about unsupported linear filtering

Most of the samples, and parts of the framework, just build samplers using
linear filtering and mipmap-modes. This is only guaranteed to be supported
for a handful of formats. On my hardware this results in a large number of
validation errors.

This change adds some helper functions to allow the filtering modes that
are preferred to be checked against the format that will be used and
drop back to nearest filtering if required.

* Fix copyright headers

* Fix commit check after rebase