Set VkAccelerationStructureGeometryTrianglesDataKHR::maxVertex to be maximum vertex index (#1097)
VkAccelerationStructureGeometryTrianglesDataKHR::maxVertex is the maximum vertex index in the VkAccelerationStructureGeometryTrianglesDataKHR::vertexData buffer i.e. the spec states "maxVertex is the number of vertices in vertexData minus one".

Currently the maxVertex value in the ray_queries and ray_tracing_extended is being set to the number of vertices, so this patch fixes them.  It's not caused any problems so far as both examples use an index buffer - but it's better to be technically correct!