Don't reopen redirected standard handles
If the calling process has redirected a standard handle (e.g. `vulkan_samples.exe sample msaa > out.log`), then reopening it will detach it from the file/pipe and connect it to the console instead.
We don't want to do that.

Instead, the type of the file should be tested before attaching the parent console.
If it's already a disk file or a pipe, then we want to leave it alone.